Conflicts:
	Ultraviolet/src/framework/HW.java
	Ultraviolet/src/framework/Teleop.java

+++ Ultraviolet/src/driver/IRSensor.java
@@ -39,6 +39,6 @@ public class IRSensor extends AnalogChannel {
         if(Math.abs(distance) > 50)
             return 3847;
 
-        return (int)distance;
+        return distance;
     }
 }
